requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn
streamlit==1.36.0
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
portalocker==2.10.1
openai==1.45.0
httpx==0.27.0
numpy